paginate_queryset, update models , certificate , subcriptions

+++ course/courses/views.py
@@ -9,16 +9,21 @@ from .services import (
     get_all_courses,
     get_course_by_id
 )
+from utils.permissions import RolePermissionFactory
+from .serializers import CourseSerializer
+from utils.pagination import paginate_queryset
 
 class CourseListView(APIView):
     def get(self, request):
         try:
             courses = get_all_courses()
-            return Response(courses, status=status.HTTP_200_OK)
+            return paginate_queryset(courses, request, CourseSerializer)
         except ValidationError as e:
             return Response({"errors": str(e)}, status=status.HTTP_404_NOT_FOUND)
 
     def post(self, request):
+        self.permission_classes = [RolePermissionFactory(['admin', 'instructor'])]
+        self.check_permissions(request)
         try:
             course = create_course(request.data)
             return Response(course, status=status.HTTP_201_CREATED)
@@ -26,6 +31,8 @@ class CourseListView(APIView):
             return Response({"errors": str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
     def patch(self, request, course_id):
+        self.permission_classes = [RolePermissionFactory(['admin', 'instructor'])]
+        self.check_permissions(request)
         try:
             course = update_course(course_id, request.data)
             return Response(course, status=status.HTTP_200_OK)
@@ -33,6 +40,8 @@ class CourseListView(APIView):
             return Response({"errors": str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, course_id):
+        self.permission_classes = [RolePermissionFactory(['admin', 'instructor'])]
+        self.check_permissions(request)
         try:
             result = delete_course(course_id)
             return Response(result, status=status.HTTP_200_OK)
@@ -42,7 +51,11 @@ class CourseListView(APIView):
 class CourseDetailView(APIView):
     def get(self, request, course_id):
         try:
-            course = get_course_by_id(course_id)
+            user = getattr(request, 'user', None)
+            # user might be AnonymousUser if no token; pass None in that case
+            if user and not hasattr(user, 'id'):
+                user = None
+            course = get_course_by_id(course_id, user=user)
             return Response(course, status=status.HTTP_200_OK)
         except ValidationError as e:
             return Response({"errors": str(e)}, status=status.HTTP_404_NOT_FOUND)
